converter/requirements.txt
pandas
requests
streamlit
opencv-python
pdf2image
openai
python-dotenv